+++ extras/wmmtestgen.py
@@ -106,7 +106,7 @@ def main(header_to_test,modelnames,dates,heights,lats,lons,bns,bes,bds,bhs,bfs,b
         margin_deg(float or int): Acceptable error in bincs and bdecs (deg)"""
     xs,ys,zs=pm.geodetic2ecef(lats,lons,heights*1000)
     bxs,bys,bzs=pm.enu2uvw(bes*1E-9,bns*1E-9,-bds*1E-9,lats,lons)
-    file_name=header_to_test[:-4]+'_test.cpp'
+    file_name='geomag_test.cpp'
     with open(file_name,'w') as f:
         outstr="""
 // %s Generated by python script wmmtestgen.py
@@ -150,7 +150,7 @@ TEST_CASE( "geomag test %s of %s model", "[GeoMag]" ) {
         f.write(outstr)
 
 if __name__ == '__main__':
-    header_to_test='geomag.hpp'
+    header_to_test='src/geomag.hpp'
     dates=([2015]*6+[2017.5]*6)*2 + [2020]*6 + [2022.5]*6
     modelnames= ['WMM2015']*12 + ['WMM2015v2']*12 + ['WMM2020']*12
     heights=([0]*3+[100]*3+[0]*3+[100]*3)*3
