+++ library.properties
@@ -1,0 +1,9 @@
+name=XYZgeomag
+version=2.0.0
+author=Nathan Zimmerberg <nhz2@cornell.edu>
+maintainer=Nathan Zimmerberg <nhz2@cornell.edu>
+sentence=Calculate the magnetic field on earth using the World Magnetic Model(WMM).
+paragraph=Given geocentric cartesian coordinates, the main function geomag::GeoMag calculates the magnetic field around earth in the International Terrestrial Reference System(ITRS) and uses units of decimal year, meter, and tesla.
+category=Data Processing
+url=https://github.com/nhz2/XYZgeomag
+architectures=*
