+++ src/geomag.hpp
@@ -23,7 +23,7 @@ SOFTWARE.
 */
 
 
-// ../geomag.hpp Generated by python script wmmcodeupdate.py
+// ../src/geomag.hpp Generated by python script wmmcodeupdate.py
 /** \file
  * \author Nathan Zimmerberg (nhz2@cornell.edu)
  * \date 07 AUG 2021
